# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API ini menyediakan serangkaian endpoint untuk mengelola users, accounts dan transactions dalam sistem. Fungsionalitas yang ditawarkan meliputi post,get,put,dan delete. API ini dirancang untuk mendukung aplikasi berbasis web dan seluler, memastikan pengalaman pengguna yang aman dan efisien.  \"berjalan di localhost:3000\""
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Server Localhost"
    }
  ],
...
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Tambah User",
        "operationId": "addUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User berhasil ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email atau identitas sudah dipakai user lain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUser400"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Dapatkan Daftar User",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Berhasil menampilkan users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Dapatkan Detail User",
        "operationId": "getUserById",
        "responses": {
          "200": {
            "description": "Berhasil menampilkan detail user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tidak ada user dengan ID tersebut",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUser404"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "operationId": "updateUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email atau identitas sudah dipakai user lain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUser400"
                }
              }
            }
          },
          "404": {
            "description": "Tidak ada user dengan ID tersebut",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUser404"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Hapus User",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": "User berhasil dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tidak ada user dengan ID tersebut",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUser404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthHeader": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gunakan header `Authorization` dengan format `Bearer <token>` untuk autentikasi anda bisa memasukkan token di sini yang sudah di buat pada melalui login.\n"
      }
    },
    "schemas": {
...
      "NewUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lewis user"
          },
          "email": {
            "type": "string",
            "example": "lewisuser@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "lewis7champions"
          },
          "identity_type": {
            "type": "string",
            "example": "KTP"
          },
          "identity_number": {
            "type": "string",
            "example": "0777999"
          },
          "address": {
            "type": "string",
            "example": "Jl. Manchester No. 14"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "identity_type",
          "identity_number",
          "address"
        ]
      },
      "NewUpdateUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lewis update user"
          },
          "email": {
            "type": "string",
            "example": "lewisupdateuser@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "lewis7champions"
          },
          "identity_type": {
            "type": "string",
            "example": "KTP"
          },
          "identity_number": {
            "type": "string",
            "example": "0777999"
          },
          "address": {
            "type": "string",
            "example": "Jl. Manchester No. 14"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "identity_type",
          "identity_number",
          "address"
        ]
      },
      "AddUserResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 201
          },
          "message": {
            "type": "string",
            "example": "user berhasil ditambahkan"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 2
              },
              "name": {
                "type": "string",
                "example": "Lewis user"
              },
              "email": {
                "type": "string",
                "example": "lewisuser@gmail.com"
              },
              "password": {
                    "type": "string",
                    "example": "$2b$10$c.54dFCQ0fgP7XZvYlh56OF90EIZlOpYPRmDsXk2zcM89o1NjgZWW"
              },
              "profile": {
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 2
                  },
                  "user_id": {
                      "type": "integer",
                      "example": 2
                    },
                  "identity_type": {
                    "type": "string",
                    "example": "KTP"
                  },
                  "identity_number": {
                    "type": "string",
                    "example": "0777999"
                  },
                  "address": {
                    "type": "string",
                    "example": "Jl. Manchester No. 14"
                  }
                }
              }
            }
          }
        }
      },
      "GetUsersResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "berhasil menampilkan users"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 2
                },
                "name": {
                  "type": "string",
                  "example": "Lewis user"
                },
                "email": {
                  "type": "string",
                  "example": "lewisuser@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "$2b$10$c.54dFCQ0fgP7XZvYlh56OF90EIZlOpYPRmDsXk2zcM89o1NjgZWW"
                }
              }
            }
          }
        }
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "berhasil menampilkan detail user"
          },
          "data": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 2
              },
              "name": {
                "type": "string",
                "example": "Lewis user"
              },
              "email": {
                "type": "string",
                "example": "lewisuser@gmail.com"
              },
              "password": {
                "type": "string",
                "example": "$2b$10$c.54dFCQ0fgP7XZvYlh56OF90EIZlOpYPRmDsXk2zcM89o1NjgZWW"
              },
              "profile": {
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 2
                  },
                  "user_id": {
                      "type": "integer",
                      "example": 1
                  },
                  "identity_type": {
                    "type": "string",
                    "example": "KTP"
                  },
                  "identity_number": {
                    "type": "string",
                    "example": "0777999"
                  },
                  "address": {
                    "type": "string",
                    "example": "Jl. Manchester No. 14"
                  }
                }
              }
            }
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "user berhasil diupdate"
          },
          "data": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "Lewis update user"
              },
              "email": {
                "type": "string",
                "example": "lewisupdateuser@gmail.com"
              },
              "password": {
                  "type": "string",
                  "example": "$2b$10$c.54dFCQ0fgP7XZvYlh56OF90EIZlOpYPRmDsXk2zcM89o1NjgZWW"
              },
              "profile": {
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 2
                  },
                  "user_id": {
                      "type": "integer",
                      "example": 1
                  },
                  "identity_type": {
                    "type": "string",
                    "example": "KTP"
                  },
                  "identity_number": {
                    "type": "string",
                    "example": "0777999"
                  },
                  "address": {
                    "type": "string",
                    "example": "Jl. Manchester No. 14"
                  }
                }
              }
            }
          }
        }
      },
      "DeleteUserResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "user berhasil dihapus"
          }
        }
      },
      "ErrorUser400": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "email atau identity_number sudah dipakai user lain"
          }
        }
      },
      "ErrorUser404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "tidak ada user"
          }
        }
      },
...
    }
  }
}